Move PredictableDict to the core `types`
